V2RaycSpider1225/requirements.txt
pytz>=2020.5
pyyaml>=5.4.1
flask>=1.1.2
easygui>=0.98.2
pyperclip>=1.8.1
redis>=3.5.3
requests>=2.25.1
bs4>=0.0.1
beautifulsoup4>=4.9.3
loguru>=0.5.3
selenium>=3.141.0
pillow>=8.3.1
gevent>=21.1.2
tqdm>=4.56.0
paramiko>=2.7.2
urllib3==1.25.11
apscheduler>=3.7.0
pymysql>=1.0.2
faker>=5.8.0
colorama>=0.4.4
pysimplegui>=4.46.0
retry>=0.9.2